+++ app/assets/stylesheets/application.scss
@@ -33,6 +33,7 @@
 @import "organisms/pagination";
 @import "organisms/provider-list";
 @import "organisms/provider-details";
+@import "organisms/welcome";
 @import "templates/about-us";
 @import "templates/consent-form";
 @import "templates/dependents-index";
