+++ app/assets/stylesheets/atoms/_variables.scss
@@ -7,3 +7,4 @@ $font-gyr:         	    "Helvetica Neue", -apple-system, BlinkMacSystemFont,
 $color-green-money: #C6F9EC;
 $color-orange-light: #FFEAD2;
 $color-orange-medium: #FFD29C;
+$color-orange-dark: #D73D02;
