+++ app/assets/stylesheets/organisms/_welcome.scss
@@ -1,0 +1,53 @@
+.welcome {
+  a {
+    text-decoration: none;
+  }
+
+  .link-box {
+    border: 0.1rem solid #A0A0A0;
+    border-radius: 1rem;
+    padding: 2rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin-bottom: 2.5rem;
+
+    &:hover {
+      background: #F7F5F4;
+    }
+
+    .link-box--content {
+      display: flex;
+      align-items: center;
+
+      h2 {
+        font-size: 1.9rem;
+        font-weight: bold;
+        margin-bottom: 1rem;
+        line-height: 2.5rem;
+      }
+
+      p {
+        font-size: 1.6rem;
+        color: #000000;
+        margin-bottom: 0;
+      }
+
+      h2.file-with-help {
+        color: $color-teal;
+      }
+
+      h2.ask-a-question {
+        color: $color-orange-dark;
+      }
+
+      img {
+        margin-right: 3.6rem;
+      }
+    }
+
+    .arrow {
+      margin-left: 1.5rem;
+    }
+  }
+}
