--- app/controllers/questions/feelings_controller.rb
@@ -1,24 +1,0 @@
-module Questions
-  class FeelingsController < QuestionsController
-    layout "question"
-
-    def current_intake
-      super || Intake.new
-    end
-
-    private
-
-    def after_update_success
-      session[:intake_id] = @form.intake.id
-    end
-
-    def illustration_path; end
-
-    def form_params
-      super.merge(
-        source: current_intake.source || source,
-        referrer: current_intake.referrer || referrer,
-      )
-    end
-  end
-end
