--- app/forms/feelings_form.rb
@@ -1,7 +1,0 @@
-class FeelingsForm < QuestionsForm
-  set_attributes_for :intake, :feeling_about_taxes, :source, :referrer
-
-  def save
-    @intake.update(attributes_for(:intake))
-  end
-end
