--- app/views/questions/feelings/edit.html.erb
@@ -1,47 +1,0 @@
-<% content_for :page_title, "How are you feeling about your taxes?" %>
-
-<% content_for :form_card do %>
-  <%= form_with model: @form, url: current_path, local: true, method: "put", builder: VitaMinFormBuilder, html: { class: "form-card form__feelings-survey" } do |f| %>
-    <fieldset class="form-group">
-      <legend class="spacing-below-35">
-        <h1 class="form-question">
-          <%= content_for(:page_title) %>
-        </h1>
-      </legend>
-
-      <%= f.hidden_field :feeling_about_taxes, value: "unfilled" %>
-      <radiogroup class="input-group--block">
-        <label class="radio-button">
-          <input type="radio" value="negative" name="feelings_form[feeling_about_taxes]" id="feelings_form_feeling_about_taxes_negative"<%= @form.feeling_about_taxes == "negative" ? " checked " : "" %>/>
-          <span class="emoji-label negative-icon"></span>
-          <span class="sr-only">Sad face</span>
-        </label>
-        <label class="radio-button">
-          <input type="radio" value="neutral" name="feelings_form[feeling_about_taxes]" id="feelings_form_feeling_about_taxes_neutral"<%= @form.feeling_about_taxes == "neutral" ? " checked " : "" %>/>
-          <span class="emoji-label neutral-icon"></span>
-          <span class="sr-only">Neutral face</span>
-        </label>
-        <label class="radio-button">
-          <input type="radio" value="positive" name="feelings_form[feeling_about_taxes]" id="feelings_form_feeling_about_taxes_positive"<%= @form.feeling_about_taxes == "positive" ? " checked " : "" %>/>
-          <span class="emoji-label positive-icon"></span>
-          <span class="sr-only">Happy face</span>
-        </label>
-      </radiogroup>
-    </fieldset>
-
-    <p>
-      We know taxes can be hard and we’re here to support you every step of the way!
-    </p>
-
-    <p>
-      Note: The federal income tax filing due date has been automatically extended to <strong>July 15, 2020</strong>.
-    </p>
-
-    <button class="button button--primary button--wide" type="submit">
-      Start my taxes online
-    </button>
-    <%= link_to vita_providers_path, class: "button button--wide text--centered" do %>
-      Find a local site
-    <% end %>
-  <% end %>
-<% end %>
