+++ spec/controllers/backtaxes_controller_spec.rb
@@ -3,13 +3,12 @@ require "rails_helper"
 RSpec.describe Questions::BacktaxesController do
   render_views
 
-  let(:intake) { create :intake }
-
-  before do
-    allow(subject).to receive(:current_intake).and_return(intake)
-  end
-
   describe "#update" do
+    before do
+      session[:source] = "source_from_session"
+      session[:referrer] = "referrer_from_session"
+    end
+
     context "with valid params" do
       let(:params) do
         {
@@ -22,13 +21,19 @@ RSpec.describe Questions::BacktaxesController do
         }
       end
 
-      it "updates the intake" do
-        post :update, params: params
+      it "creates new intake with the backtaxes answers" do
+        expect {
+          post :update, params: params
+        }.to change(Intake, :count).by(1)
 
+        intake = Intake.last
+        expect(intake.source).to eq "source_from_session"
+        expect(intake.referrer).to eq "referrer_from_session"
         expect(intake.needs_help_2016).to eq "no"
         expect(intake.needs_help_2017).to eq "yes"
         expect(intake.needs_help_2018).to eq "yes"
         expect(intake.needs_help_2019).to eq "yes"
+        expect(session[:intake_id]).to eq intake.id
       end
     end
