+++ spec/controllers/documents_controller_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe DocumentsController, type: :controller do
         delete :destroy, params: { id: document.id }
       end.not_to change(Document, :count)
 
-      expect(response).to redirect_to(feelings_questions_path)
+      expect(response).to redirect_to(welcome_questions_path)
     end
 
     context "with an authenticated user" do
