+++ spec/features/web_intake/new_joint_filers_spec.rb
@@ -14,11 +14,13 @@ RSpec.feature "Web Intake Joint Filers" do
   end
 
   scenario "new client filing joint taxes with spouse and dependents", :js do
-    # Feelings
-    visit "/questions/feelings"
-    expect(page).to have_selector("h1", text: "How are you feeling about your taxes?")
-    choose "Happy face"
-    click_on "Start my taxes online"
+    # Home
+    visit "/"
+    find("#firstCta").click
+
+    # Welcome
+    expect(page).to have_selector("h1", text: "Welcome! How can we help you?")
+    click_on "File taxes with help"
 
     # File With Help
     expect(current_path).to eq(file_with_help_questions_path)
